refactor(app): move routes.go and sentry middleware to internal packages - move routes.go to internal/app/routes.go - move middleware.go to internal/middleware/sentry_middleware.go
