+++ internal/middleware/sentry_middleware.go
@@ -1,4 +1,4 @@
-package main
+package middleware
 
 import (
 	"net/http"
@@ -7,7 +7,7 @@ import (
 	"github.com/getsentry/sentry-go/http"
 )
 
-func sentryMiddleware(next http.Handler) http.Handler {
+func SentryMiddleware(next http.Handler) http.Handler {
 	sentryHandler := sentryhttp.New(sentryhttp.Options{
 		Repanic:         true,
 		WaitForDelivery: true,
